requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
Flask~=3.1.0
git+https://github.com/esoadamo/Python-SQLiteDB.git
